Added debug render via ImGui

+++ src/Include/xrRender/DebugRender.h
@@ -25,7 +25,7 @@ public:
 public:
 	virtual			~IDebugRender		()	{}
 	virtual void	Render				() = 0;
-	virtual void	add_lines			(Fvector const *vertices, u32 const &vertex_count, u16 const *pairs, u32 const &pair_count, u32 const &color) = 0;
+	virtual void	add_lines			(Fvector const *vertices, u32 const &vertex_count, u32 const *pairs, u32 const &pair_count, u32 const &color) = 0;
 	
 	// routed to RCache
 	virtual void	NextSceneMode		() = 0;
